Refine platform sending behavior
