+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 所有重要变更都会记录在这里。
 
+## [0.6.9] - 2026-06-07
+
+### Changed
+
+- 飞书/Lark 不再进入合并调度；逐账号普通发送时优先使用原生 `post` 将正文和本地图片放在同一条消息中，失败时降级到原有 `text` 正文加普通媒体附件路径。
+- `merge_tweet_threshold` 收敛为 QQ/`aiocqhttp` 专用合并转发阈值；飞书/Lark、Telegram、微信 OC 和其他平台不再随阈值进入合并调度，始终逐账号普通发送。
+- README、配置提示和插件元数据同步说明飞书/Lark 不支持 QQ 式合并转发，飞书单账号消息仅作为平台原生展示优化。
+
 ## [0.6.8] - 2026-06-07
 
 ### Added
