+++ _conf_schema.json
@@ -3,16 +3,7 @@
     "description": "Nitter 实例列表",
     "type": "list",
     "default": [
-      "https://nitter.catsarch.com",
-      "https://nitter.tiekoetter.com",
-      "https://nuku.trabun.org",
-      "https://nitter.privacyredirect.com",
-      "https://xcancel.com",
-      "https://nitter.net",
-      "https://nitter.space",
-      "https://lightbrd.com",
-      "https://nitter.kareem.one",
-      "https://nitter.poast.org"
+      "https://nitter.net"
     ],
     "hint": "按顺序尝试 /<username>/rss。公共实例不适合高频抓取，长期使用建议改为自建实例。"
   },
@@ -209,10 +200,10 @@
     "hint": "关闭时插件重启后不会立刻检查新推文，而是等到下一个间隔或每日定点；开启后恢复旧行为。"
   },
   "merge_tweet_threshold": {
-    "description": "合并推送触发条数",
+    "description": "QQ 合并推送触发条数",
     "type": "int",
     "default": 2,
-    "hint": "本次发送的新推文总数达到该值时启用合并发送；0 表示关闭，1 表示有新推文就合并，默认 2 表示 2 条及以上才合并。OneBot v11/aiocqhttp 达到阈值时使用 Node/Nodes 合并转发；飞书/Lark 达到阈值时使用原生 post 富文本发送合并正文，未达到阈值时使用原生 text 正文；其他平台使用普通消息链。范围 0-20。"
+    "hint": "QQ/aiocqhttp 目标在本次发送的新推文总数达到该值时启用 OneBot v11 Node/Nodes 合并转发；0 表示关闭，1 表示有新推文就合并，默认 2 表示 2 条及以上才合并。飞书/Lark、Telegram、微信 OC 和其他平台不受该阈值影响，始终使用普通逐账号发送。范围 0-20。"
   },
   "watch_users": {
     "description": "关注账号列表",
