+++ metadata.yaml
@@ -1,8 +1,8 @@
 name: astrbot_plugin_nitter_tweets
 display_name: Nitter 推文记录
 desc: "[多平台接入开发中] Nitter 推文记录 - 支持 QQ (aiocqhttp)、Telegram、Feishu (Lark) 以及微信 OC (weixin_oc)；通过 Nitter RSS 获取公开推文和媒体，支持手动查询与定时推送。"
-version: 0.6.8
-description: 通过可配置的 Nitter 实例轮询指定用户 RSS，将最近公开推文发送到聊天会话；支持手动查询和独立的 Nitter 镜像站测试命令；OneBot v11 可按推文数量阈值使用 Node/Nodes 合并转发，飞书/Lark 达到合并阈值时使用原生 post 富文本发送合并正文，未达到阈值时使用原生 text 发送正文，其他平台使用普通消息链。可通过 xdown 解析并发送图片附件，视频/GIF 附件默认关闭并保留原帖链接；支持按间隔或每日定点检查关注账号并推送新推文，可清理重复订阅作者；可调用 AstrBot 已配置大模型翻译非中文推文，并按概率追加 AI 评论与识图描述。
+version: 0.6.9
+description: 通过可配置的 Nitter 实例轮询指定用户 RSS，将最近公开推文发送到聊天会话；支持手动查询和独立的 Nitter 镜像站测试命令；QQ/aiocqhttp 可按推文数量阈值使用 OneBot v11 Node/Nodes 合并转发，飞书/Lark、Telegram、微信 OC 和其他平台始终普通逐账号发送；飞书逐账号发送时会优先使用原生 post 同框发送正文和图片。可通过 xdown 解析并发送图片附件，视频/GIF 附件默认关闭并保留原帖链接；支持按间隔或每日定点检查关注账号并推送新推文，可清理重复订阅作者；可调用 AstrBot 已配置大模型翻译非中文推文，并按概率追加 AI 评论与识图描述。
 author: shitianyaa
 astrbot_version: ">=4.16.0"
 repo: https://github.com/shitianyaa/astrbot_plugin_nitter_tweets
